+++ docs/source/conf.py
@@ -1,5 +1,9 @@
 import os
 import sys
+import inspect
+import logging
+import operator
+
 
 # skrl library
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..", "..")))
@@ -10,7 +14,7 @@ import skrl
 
 # project information
 project = "skrl"
-copyright = "2021-2025, Toni-SM"
+copyright = "2021-2026, Toni-SM"
 author = "Toni-SM"
 
 if skrl.__version__ != "unknown":
@@ -22,11 +26,13 @@ master_doc = "index"
 
 # general configuration
 extensions = [
+    "autodocsumm",
     "sphinx.ext.duration",
     "sphinx.ext.doctest",
     "sphinx.ext.autodoc",
     "sphinx.ext.autosummary",
     "sphinx.ext.intersphinx",
+    "sphinx.ext.linkcode",
     "sphinx_tabs.tabs",
     "sphinx_copybutton",
     "notfound.extension",
@@ -121,20 +127,26 @@ epub_show_urls = "footnote"
 
 # autodoc ext
 autodoc_member_order = "groupwise"
-autoclass_content = "init"
+autodoc_default_options = {
+    "autosummary": True,
+}
 autodoc_mock_imports = [
+    "flax",
     "gym",
     "gymnasium",
-    "torch",
     "jax",
     "jaxlib",
-    "flax",
+    "mujoco",
     "optax",
-    "warp",
-    "tensorboard",
-    "tqdm",
     "packaging",
+    "tensorboard",
+    "torch",
+    "tqdm",
+    "warp",
 ]
+autoclass_content = "init"
+autosummary_generate = True
+autosummary_generate_overwrite = False
 
 # copybutton ext
 copybutton_prompt_text = r">>> |\.\.\. "
@@ -151,16 +163,43 @@ notfound_context = {
 """,
 }
 
+
+# linkcode ext
+def linkcode_resolve(domain, info):
+    if domain != "py":
+        return None
+    if not info["module"]:
+        return None
+    if not info["fullname"]:
+        return None
+
+    try:
+        mod = sys.modules.get(info["module"])
+        obj = operator.attrgetter(info["fullname"])(mod)
+        if isinstance(obj, property):
+            obj = obj.fget
+        filename = inspect.getsourcefile(obj)
+        source, linenum = inspect.getsourcelines(obj)
+    except Exception:
+        return None
+
+    github_version = os.environ.get("GITHUB_REF_NAME") or os.environ.get("CI_COMMIT_REF_NAME") or "develop"
+    filename = os.path.relpath(filename, start=os.path.dirname(skrl.__file__))
+    lines = f"#L{linenum}-L{linenum + len(source)}" if linenum else ""
+
+    return f"https://github.com/Toni-SM/skrl/blob/{github_version}/skrl/{filename}{lines}"
+
+
 # suppress warning messages
 suppress_warnings = [
     "ref.python",  # more than one target found for cross-reference
 ]
 
-# hack to suppress 'WARNING: duplicate object description... use :no-index:'
-import logging
 
+# hack to suppress 'WARNING: duplicate object description... use :no-index:'
 class DuplicateObjectDescriptionFilter(logging.Filter):
     def filter(self, record):
         return "duplicate object description" not in record.getMessage()
+
 
 logging.getLogger("sphinx.sphinx.domains.python").addFilter(DuplicateObjectDescriptionFilter())