+++ skrl/trainers/jax/step.py
@@ -134,7 +134,7 @@ class StepTrainer(Trainer):
                     agent.track_data("Stats / Env stepping time (ms)", elapsed_time_ms)
 
             # render the environments
-            if not self.cfg.headless:
+            if not self.cfg.headless and not timestep % self.cfg.render_interval:
                 self.env.render()
 
             # record the environments' transitions
@@ -252,7 +252,7 @@ class StepTrainer(Trainer):
                     agent.track_data("Stats / Env stepping time (ms)", elapsed_time_ms)
 
             # render the environments
-            if not self.cfg.headless:
+            if not self.cfg.headless and not timestep % self.cfg.render_interval:
                 self.env.render()
 
             # write data to TensorBoard