+++ tests/envs/wrappers/jax/test_gymnasium_envs.py
@@ -6,7 +6,8 @@ import gymnasium as gym
 import jax
 import jax.numpy as jnp
 
-from skrl.envs.wrappers.jax import GymnasiumWrapper, wrap_env
+from skrl.envs.wrappers.jax import wrap_env
+from skrl.envs.wrappers.jax.gymnasium_envs import GymnasiumWrapper
 
 
 def test_env(capsys: pytest.CaptureFixture):