+++ tests/envs/wrappers/jax/test_isaaclab_envs.py
@@ -8,8 +8,8 @@ import jax.numpy as jnp
 import numpy as np
 import torch
 
-from skrl import config
-from skrl.envs.wrappers.jax import IsaacLabMultiAgentWrapper, IsaacLabWrapper, wrap_env
+from skrl.envs.wrappers.jax import wrap_env
+from skrl.envs.wrappers.jax.isaaclab_envs import IsaacLabMultiAgentWrapper, IsaacLabWrapper
 
 
 class IsaacLabEnv(gym.Env):
@@ -146,8 +146,7 @@ def test_env(capsys: pytest.CaptureFixture, num_states: int):
     action = jnp.ones((num_envs, 1))
 
     # check wrapper definition
-    with pytest.raises(AttributeError):
-        assert isinstance(wrap_env(None, "isaaclab-single-agent"), IsaacLabWrapper)
+    assert isinstance(wrap_env(None, "isaaclab-single-agent"), IsaacLabWrapper)
 
     # load wrap the environment
     original_env = IsaacLabEnv(num_states)
@@ -206,8 +205,7 @@ def test_multi_agent_env(capsys: pytest.CaptureFixture, num_states: int):
     action = {f"agent_{i}": jnp.ones((num_envs, i + 10)) for i in range(num_agents)}
 
     # check wrapper definition
-    with pytest.raises(AttributeError):
-        assert isinstance(wrap_env(None, "isaaclab-multi-agent"), IsaacLabMultiAgentWrapper)
+    assert isinstance(wrap_env(None, "isaaclab-multi-agent"), IsaacLabMultiAgentWrapper)
 
     # load wrap the environment
     original_env = IsaacLabMultiAgentEnv(num_states)