+++ tests/trainers/torch/test_step.py
@@ -23,6 +23,7 @@ from ...utilities import (
     # trainer config
     timesteps=st.integers(min_value=1, max_value=50),
     headless=st.booleans(),
+    render_interval=st.integers(min_value=1, max_value=5),
     disable_progressbar=st.booleans(),
     close_environment_at_exit=st.booleans(),
     stochastic_evaluation=st.booleans(),
@@ -42,6 +43,7 @@ def test_non_simultaneous_trainer_single_agent(
     # trainer config
     timesteps,
     headless,
+    render_interval,
     disable_progressbar,
     close_environment_at_exit,
     stochastic_evaluation,
@@ -80,6 +82,7 @@ def test_non_simultaneous_trainer_single_agent(
     cfg = {
         "timesteps": timesteps,
         "headless": headless,
+        "render_interval": render_interval,
         "disable_progressbar": disable_progressbar,
         "close_environment_at_exit": close_environment_at_exit,
         "environment_info": "episode",
@@ -102,6 +105,7 @@ def test_non_simultaneous_trainer_single_agent(
     # trainer config
     timesteps=st.integers(min_value=1, max_value=50),
     headless=st.booleans(),
+    render_interval=st.integers(min_value=1, max_value=5),
     disable_progressbar=st.booleans(),
     close_environment_at_exit=st.booleans(),
     stochastic_evaluation=st.booleans(),
@@ -122,6 +126,7 @@ def test_simultaneous_trainer_single_agent(
     # trainer config
     timesteps,
     headless,
+    render_interval,
     disable_progressbar,
     close_environment_at_exit,
     stochastic_evaluation,
@@ -170,6 +175,7 @@ def test_simultaneous_trainer_single_agent(
     cfg = {
         "timesteps": timesteps,
         "headless": headless,
+        "render_interval": render_interval,
         "disable_progressbar": disable_progressbar,
         "close_environment_at_exit": close_environment_at_exit,
         "environment_info": "episode",
@@ -192,6 +198,7 @@ def test_simultaneous_trainer_single_agent(
     # trainer config
     timesteps=st.integers(min_value=1, max_value=50),
     headless=st.booleans(),
+    render_interval=st.integers(min_value=1, max_value=5),
     disable_progressbar=st.booleans(),
     close_environment_at_exit=st.booleans(),
     stochastic_evaluation=st.booleans(),
@@ -212,6 +219,7 @@ def test_non_simultaneous_trainer_multi_agent(
     # trainer config
     timesteps,
     headless,
+    render_interval,
     disable_progressbar,
     close_environment_at_exit,
     stochastic_evaluation,
@@ -259,6 +267,7 @@ def test_non_simultaneous_trainer_multi_agent(
     cfg = {
         "timesteps": timesteps,
         "headless": headless,
+        "render_interval": render_interval,
         "disable_progressbar": disable_progressbar,
         "close_environment_at_exit": close_environment_at_exit,
         "environment_info": "episode",
@@ -282,6 +291,7 @@ def test_non_simultaneous_trainer_multi_agent(
     # trainer config
     timesteps=st.integers(min_value=1, max_value=50),
     headless=st.booleans(),
+    render_interval=st.integers(min_value=1, max_value=5),
     disable_progressbar=st.booleans(),
     close_environment_at_exit=st.booleans(),
     stochastic_evaluation=st.booleans(),
@@ -303,6 +313,7 @@ def test_simultaneous_trainer_multi_agent(
     # trainer config
     timesteps,
     headless,
+    render_interval,
     disable_progressbar,
     close_environment_at_exit,
     stochastic_evaluation,
@@ -360,6 +371,7 @@ def test_simultaneous_trainer_multi_agent(
     cfg = {
         "timesteps": timesteps,
         "headless": headless,
+        "render_interval": render_interval,
         "disable_progressbar": disable_progressbar,
         "close_environment_at_exit": close_environment_at_exit,
         "environment_info": "episode",